Short pages don't scroll off the bottom of the page.
The .wrapper div has height 100%, which ends up pushing the menu off
the bottom of the page, since the negative margin at the bottom wasn't
sufficient to pull the menu (and the space below it) up far enough
(this negative margin was increased). Also, there was padding at the
top of .wrapper, which pushes the whole page down, so this was moved
to be at the top of #header inside .wrapper.